pyproject.toml
[project]
name = "jammie-disks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.135.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
]
